+++ bot.py
@@ -284,7 +284,7 @@ class ModmailBot(commands.Bot):
                     loop.run_until_complete(loop.shutdown_asyncgens())
             finally:
                 logger.info("Closing the event loop.")
-                loop.close()
+                # loop.close()
 
         if not future.cancelled():
             try: