+++ cogs/modmail.py
@@ -1038,7 +1038,14 @@ class Modmail(commands.Cog):
         if isinstance(category, str):
             if "silent" in category or "silently" in category:
                 silent = True
-            category = None
+                category = category.strip("silently").strip("silent").strip()
+                try:
+                    category = await SimilarCategoryConverter().convert(ctx, category)  # attempt to find a category again
+                except commands.BadArgument:
+                    category = None
+
+            if isinstance(category, str):
+                category = None
 
         if user.bot:
             embed = discord.Embed(color=self.bot.error_color, description="Cannot start a thread with a bot.")